+++ CodeIgniter3-V3.1.13/application/models/Users_Model.php
@@ -5,4 +5,28 @@ class Users_Model extends CI_Model
 	{
 		return $this->db->get('users')->result_array();
 	}
+
+	public function findUser($id)
+	{
+		$findUser = $this->db->get_where('users', ['id' => $id])->row();
+		return $findUser;
+	}
+
+	public function deleteUser($id)
+	{
+		$this->db->delete('users', ['id' => $id]);
+		return $this->db->affected_rows();
+	}
+
+	public function createUser($data)
+	{
+		$this->db->insert('users', $data);
+		return $this->db->affected_rows();
+	}
+
+	public function updateUser($data, $id)
+	{
+		$this->db->update('users', $data, ['id' => $id]);
+		return $this->db->affected_rows();
+	}
 }