+++ CHANGELOG.md
@@ -1,5 +1,15 @@
 # Release Notes
 
+## v1.0.4 (2023-08-27)
+
+- update users
+
+## v1.0.3 (2023-08-27)
+
+- find users
+- delete users
+- create users
+
 ## v1.0.2 (2023-08-27)
 
 - response json